+++ cmake/BLTMacros.cmake
@@ -184,7 +184,13 @@ macro(blt_add_target_compile_flags)
     # Only add the flag if it is not empty
     string(STRIP "${arg_FLAGS}" _strippedFlags)
     if(NOT "${_strippedFlags}" STREQUAL "")
-        target_compile_options(${arg_TO} ${_scope} ${_strippedFlags})
+        get_target_property(_target_type ${arg_TO} TYPE)
+        if (("${_target_type}" STREQUAL "INTERFACE_LIBRARY") AND (${CMAKE_VERSION} VERSION_LESS "3.11.0"))
+            set_property(TARGET ${arg_NAME} APPEND PROPERTY
+                         INTERFACE_COMPILE_OPTIONS ${_strippedFlags})
+        else()
+            target_compile_options(${arg_TO} ${_scope} ${_strippedFlags})
+        endif()
     endif()
 
     unset(_strippedFlags)
@@ -270,16 +276,30 @@ macro(blt_add_target_link_flags)
             endif()
         else()
             # In CMake <= 3.12, there is no target_link_flags or target_link_options command
-            get_target_property(_link_flags ${arg_TO} LINK_FLAGS)
-            if(NOT _link_flags)
-                set(_link_flags "")
-            endif()
-            set(_link_flags "${_flags} ${_link_flags}")
+            get_target_property(_target_type ${arg_TO} TYPE)
+            if ("${_target_type}" STREQUAL "INTERFACE_LIBRARY")
+                # If it's an interface library, we add the flag via link_libraries
+                if(${CMAKE_VERSION} VERSION_GREATER_EQUAL "3.11.0")
+                    target_link_libraries(${arg_TO} INTERFACE ${_flags})
+                else()
+                    set_property(TARGET ${arg_NAME} APPEND PROPERTY
+                                 INTERFACE_LINK_LIBRARIES ${_flags})
+                endif()
+            else()
+                get_target_property(_link_flags ${arg_TO} LINK_FLAGS)
+                # Append to existing flags
+                if(NOT _link_flags)
+                    set(_link_flags "")
+                endif()
+                set(_link_flags "${_flags} ${_link_flags}")
 
-            # Convert from a CMake ;-list to a string
-            string (REPLACE ";" " " _link_flags_str "${_link_flags}")
-            set_target_properties(${arg_TO}
-                                  PROPERTIES LINK_FLAGS "${_link_flags_str}")
+                # Convert from a CMake ;-list to a string
+                string (REPLACE ";" " " _link_flags_str "${_link_flags}")
+
+                set_target_properties(${arg_TO}
+                                      PROPERTIES LINK_FLAGS "${_link_flags_str}")
+            endif()
+
         endif()
     endif()
 
@@ -379,6 +399,175 @@ macro(blt_register_library)
     endif()
 
 endmacro(blt_register_library)
+
+##------------------------------------------------------------------------------
+## blt_patch_target( NAME <targetname>
+##                   DEPENDS_ON [dep1 [dep2 ...]]
+##                   INCLUDES [include1 [include2 ...]]
+##                   LIBRARIES [lib1 [lib2 ...]]
+##                   TREAT_INCLUDES_AS_SYSTEM [ON|OFF]
+##                   FORTRAN_MODULES [ path1 [ path2 ..]]
+##                   COMPILE_FLAGS [ flag1 [ flag2 ..]]
+##                   LINK_FLAGS [ flag1 [ flag2 ..]]
+##                   DEFINES [def1 [def2 ...]] )
+##
+## Modifies an existing CMake target - sets PUBLIC visibility except for INTERFACE
+## libraries, which use INTERFACE visibility
+##------------------------------------------------------------------------------
+macro(blt_patch_target)
+    set(singleValueArgs NAME TREAT_INCLUDES_AS_SYSTEM)
+    set(multiValueArgs INCLUDES 
+                       DEPENDS_ON
+                       LIBRARIES
+                       FORTRAN_MODULES
+                       COMPILE_FLAGS
+                       LINK_FLAGS
+                       DEFINES )
+
+    ## parse the arguments
+    cmake_parse_arguments(arg
+        "${options}" "${singleValueArgs}" "${multiValueArgs}" ${ARGN} )
+
+    # Input checks
+    if( "${arg_NAME}" STREQUAL "" )
+        message(FATAL_ERROR "blt_patch_target() must be called with argument NAME <name>")
+    endif()
+
+    if (NOT TARGET ${arg_NAME})
+        message(FATAL_ERROR "blt_patch_target() NAME argument must be a native CMake target")
+    endif()
+
+    # Default to public scope, unless it's an interface library
+    set(_scope PUBLIC)
+    get_target_property(_target_type ${arg_NAME} TYPE)
+    if("${_target_type}" STREQUAL "INTERFACE_LIBRARY")
+        set(_scope INTERFACE)
+    endif()
+
+    # Interface libraries were heavily restricted pre-3.11
+    set(_standard_lib_interface FALSE)
+    if((${CMAKE_VERSION} VERSION_GREATER_EQUAL "3.11.0") OR (NOT "${_target_type}" STREQUAL "INTERFACE_LIBRARY"))
+        set(_standard_lib_interface TRUE)
+    endif()
+
+    # LIBRARIES and DEPENDS_ON are kept separate in case different logic is needed for
+    # the library itself versus its dependencies
+    set(_libs_to_link "")
+    if( arg_LIBRARIES )
+        list(APPEND _libs_to_link ${arg_LIBRARIES})
+    endif()
+
+    # TODO: This won't expand BLT-registered libraries
+    if( arg_DEPENDS_ON )
+        list(APPEND _libs_to_link ${arg_DEPENDS_ON})
+    endif()
+
+    if(_standard_lib_interface)
+        target_link_libraries(${arg_NAME} ${_scope} ${_libs_to_link})
+    else()
+        set_property(TARGET ${arg_NAME} APPEND PROPERTY
+                     INTERFACE_LINK_LIBRARIES ${_libs_to_link})
+    endif()
+        
+    if( arg_INCLUDES )
+        if(_standard_lib_interface)
+            target_include_directories(${arg_NAME} ${_scope} ${arg_INCLUDES})
+        else()
+            # Interface include directories need to be set manually
+            set_property(TARGET ${arg_NAME} APPEND PROPERTY 
+                         INTERFACE_INCLUDE_DIRECTORIES ${arg_INCLUDES})
+        endif()
+    endif()
+
+    # PGI does not support -isystem
+    if( (${arg_TREAT_INCLUDES_AS_SYSTEM}) AND (NOT "${CMAKE_CXX_COMPILER_ID}" STREQUAL "PGI"))
+        get_target_property(_target_includes ${arg_NAME} INTERFACE_INCLUDE_DIRECTORIES)
+        if(_standard_lib_interface)
+            target_include_directories(${arg_NAME} SYSTEM ${_scope} ${_target_includes})
+        else()
+            set_property(TARGET ${arg_NAME} PROPERTY
+                         INTERFACE_SYSTEM_INCLUDE_DIRECTORIES ${_target_includes})
+        endif()
+    endif()
+
+    # FIXME: Is this all that's needed?
+    if( arg_FORTRAN_MODULES )
+        target_include_directories(${arg_NAME} ${_scope} ${arg_FORTRAN_MODULES})
+    endif()
+
+    if( arg_COMPILE_FLAGS )
+        blt_add_target_compile_flags(TO ${arg_NAME} 
+                                     SCOPE ${_scope}
+                                     FLAGS ${arg_COMPILE_FLAGS})
+    endif()
+    
+    if( arg_LINK_FLAGS )
+        blt_add_target_link_flags(TO ${arg_NAME} 
+                                  SCOPE ${_scope}
+                                  FLAGS ${arg_LINK_FLAGS})
+    endif()
+    
+    if( arg_DEFINES )
+        blt_add_target_definitions(TO ${arg_NAME} 
+                                   SCOPE ${_scope}
+                                   TARGET_DEFINITIONS ${arg_DEFINES})
+    endif()
+
+endmacro(blt_patch_target)
+
+##------------------------------------------------------------------------------
+## blt_import_library( NAME <libname>
+##                     LIBRARIES [lib1 [lib2 ...]]
+##                     DEPENDS_ON [dep1 [dep2 ...]]
+##                     INCLUDES [include1 [include2 ...]]
+##                     TREAT_INCLUDES_AS_SYSTEM [ON|OFF]
+##                     FORTRAN_MODULES [ path1 [ path2 ..]]
+##                     COMPILE_FLAGS [ flag1 [ flag2 ..]]
+##                     LINK_FLAGS [ flag1 [ flag2 ..]]
+##                     DEFINES [def1 [def2 ...]] 
+##                     GLOBAL [ON|OFF])
+##
+## Imports a library as a CMake target
+##------------------------------------------------------------------------------
+macro(blt_import_library)
+    set(singleValueArgs NAME TREAT_INCLUDES_AS_SYSTEM GLOBAL)
+    set(multiValueArgs LIBRARIES
+                       INCLUDES 
+                       DEPENDS_ON
+                       FORTRAN_MODULES
+                       COMPILE_FLAGS
+                       LINK_FLAGS
+                       DEFINES )
+
+    ## parse the arguments
+    cmake_parse_arguments(arg
+        "${options}" "${singleValueArgs}" "${multiValueArgs}" ${ARGN} )
+
+    # Input checks
+    if( "${arg_NAME}" STREQUAL "" )
+        message(FATAL_ERROR "blt_import_library() must be called with argument NAME <name>")
+    endif()
+
+    # Add all imported targets to a single imported interface target
+    if(${arg_GLOBAL})
+        add_library(${arg_NAME} INTERFACE IMPORTED GLOBAL)
+    else()
+        add_library(${arg_NAME} INTERFACE IMPORTED)
+    endif()
+
+    blt_patch_target(
+        NAME       ${arg_NAME}
+        LIBRARIES  ${arg_LIBRARIES}
+        DEPENDS_ON ${arg_DEPENDS_ON}
+        INCLUDES   ${arg_INCLUDES}
+        DEFINES    ${arg_DEFINES}
+        TREAT_INCLUDES_AS_SYSTEM ${arg_TREAT_INCLUDES_AS_SYSTEM}
+        FORTRAN_MODULES ${arg_FORTRAN_MODULES}
+        COMPILE_FLAGS ${arg_COMPILE_FLAGS}
+        LINK_FLAGS ${arg_LINK_FLAGS}
+        DEFINES ${arg_DEFINES}
+    )
+endmacro(blt_import_library)
 
 
 ##------------------------------------------------------------------------------
@@ -1130,8 +1319,8 @@ macro(blt_print_target_properties)
         blt_list_remove_duplicates(TO _property_list)   
 
         ## For interface targets, filter against whitelist of valid properties
-        get_property(_targetType TARGET ${arg_TARGET} PROPERTY TYPE)
-        if(${_targetType} STREQUAL "INTERFACE_LIBRARY")
+        get_property(_target_type TARGET ${arg_TARGET} PROPERTY TYPE)
+        if("${_target_type}" STREQUAL "INTERFACE_LIBRARY")
             blt_filter_list(TO _property_list
                             REGEX "^(INTERFACE_|IMPORTED_LIBNAME_|COMPATIBLE_INTERFACE_|MAP_IMPORTED_CONFIG_)|^(NAME|TYPE|EXPORT_NAME)$"
                             OPERATION "include")