+++ cmake/BLTPrivateMacros.cmake
@@ -92,25 +92,38 @@ endfunction()
 function(blt_fix_fortran_openmp_flags target_name)
 
     if (ENABLE_FORTRAN AND ENABLE_OPENMP AND BLT_OPENMP_FLAGS_DIFFER)
+        # The OpenMP interface library will have been added as a direct
+        # link dependency instead of via flags
+        get_target_property(target_link_libs ${target_name} LINK_LIBRARIES)
+        if ( target_link_libs )
+            # Since this is only called on executable targets we can safely convert
+            # from a "real" target back to a "fake" one as this is a sink vertex in
+            # the DAG.  Only the link flags need to be modified.
+            list(FIND target_link_libs "openmp" _omp_index)
+            if(${_omp_index} GREATER -1)
+                list(REMOVE_ITEM target_link_libs "openmp")
 
-        set(_property_name LINK_FLAGS)
-        if( ${CMAKE_VERSION} VERSION_GREATER_EQUAL "3.13.0" )
-            # In CMake 3.13+, LINK_FLAGS was converted to LINK_OPTIONS.
-            set(_property_name LINK_OPTIONS)
-        endif()
+                # Copy the compile flags verbatim
+                get_target_property(omp_compile_flags openmp INTERFACE_COMPILE_OPTIONS)
+                target_compile_options(${target_name} PUBLIC ${omp_compile_flags})
 
-        get_target_property(target_link_flags ${target_name} ${_property_name})
-        if ( target_link_flags )
+                # These are set through blt_add_target_link_flags which needs to use
+                # the link_libraries for interface libraries in CMake < 3.13
+                if( ${CMAKE_VERSION} VERSION_GREATER_EQUAL "3.13.0" )
+                    get_target_property(omp_link_flags openmp INTERFACE_LINK_OPTIONS)
+                else()
+                    get_target_property(omp_link_flags openmp INTERFACE_LINK_LIBRARIES)
+                endif()
+    
+                string( REPLACE "${OpenMP_CXX_FLAGS}" "${OpenMP_Fortran_FLAGS}"
+                        correct_omp_link_flags
+                        "${omp_link_flags}"
+                        )
+                list(APPEND target_link_libs "${correct_omp_link_flags}")
+                set_target_properties( ${target_name} PROPERTIES
+                                       LINK_LIBRARIES "${target_link_libs}" )
+            endif()
 
-            message(STATUS "Fixing OpenMP Flags for target[${target_name}]")
-
-            string( REPLACE "${OpenMP_CXX_FLAGS}" "${OpenMP_Fortran_FLAGS}"
-                    correct_link_flags
-                    "${target_link_flags}"
-                    )
-
-            set_target_properties( ${target_name} PROPERTIES ${_property_name}
-                                   "${correct_link_flags}" )
         endif()
 
     endif()
@@ -226,6 +239,20 @@ macro(blt_inherit_target_info)
                         ${arg_FROM} INTERFACE_COMPILE_DEFINITIONS)
     if ( _interface_defines )
         target_compile_definitions( ${arg_TO} PUBLIC ${_interface_defines})
+    endif()
+
+    if( ${CMAKE_VERSION} VERSION_GREATER_EQUAL "3.13.0" )
+        get_target_property(_interface_link_options
+                            ${arg_FROM} INTERFACE_LINK_OPTIONS)
+        if ( _interface_link_options )
+            target_link_options( ${arg_TO} PUBLIC ${_interface_link_options})
+        endif()
+    endif()
+
+    get_target_property(_interface_compile_options
+                        ${arg_FROM} INTERFACE_COMPILE_OPTIONS)
+    if ( _interface_compile_options )
+        target_compile_options( ${arg_TO} PUBLIC ${_interface_compile_options})
     endif()
 
     if ( NOT arg_OBJECT )
@@ -352,7 +379,6 @@ macro(blt_setup_target)
             blt_add_target_link_flags(TO ${arg_NAME}
                                       FLAGS ${_BLT_${uppercase_dependency}_LINK_FLAGS} )
         endif()
-
     endforeach()
 
 endmacro(blt_setup_target)